Remove video install step
