--- apps/web/app/(use-page-wrapper)/onboarding/personal/video/page.tsx
@@ -1,34 +1,0 @@
-import { _generateMetadata } from "app/_utils";
-import { cookies, headers } from "next/headers";
-import { redirect } from "next/navigation";
-
-import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
-import { APP_NAME } from "@calcom/lib/constants";
-
-import { buildLegacyRequest } from "@lib/buildLegacyCtx";
-
-import { PersonalVideoView } from "~/onboarding/personal/video/personal-video-view";
-
-export const generateMetadata = async () => {
-  return await _generateMetadata(
-    (t) => `${APP_NAME} - ${t("connect_video")}`,
-    () => "",
-    true,
-    undefined,
-    "/onboarding/personal/video"
-  );
-};
-
-const ServerPage = async () => {
-  const session = await getServerSession({ req: buildLegacyRequest(await headers(), await cookies()) });
-
-  if (!session?.user?.id) {
-    return redirect("/auth/login");
-  }
-
-  const userEmail = session.user.email || "";
-
-  return <PersonalVideoView userEmail={userEmail} />;
-};
-
-export default ServerPage;
