+++ apps/web/modules/onboarding/hooks/useSubmitPersonalOnboarding.ts
@@ -1,0 +1,77 @@
+import { useRouter } from "next/navigation";
+
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { useTelemetry } from "@calcom/lib/hooks/useTelemetry";
+import { telemetryEventTypes } from "@calcom/lib/telemetry";
+import { trpc } from "@calcom/trpc/react";
+import { showToast } from "@calcom/ui/components/toast";
+
+const DEFAULT_EVENT_TYPES = [
+  {
+    title: "15min_meeting",
+    slug: "15min",
+    length: 15,
+  },
+  {
+    title: "30min_meeting",
+    slug: "30min",
+    length: 30,
+  },
+  {
+    title: "secret_meeting",
+    slug: "secret",
+    length: 15,
+    hidden: true,
+  },
+];
+
+export const useSubmitPersonalOnboarding = () => {
+  const router = useRouter();
+  const { t } = useLocale();
+  const telemetry = useTelemetry();
+  const utils = trpc.useUtils();
+
+  const { data: eventTypes } = trpc.viewer.eventTypes.list.useQuery();
+  const createEventType = trpc.viewer.eventTypesHeavy.create.useMutation();
+
+  const mutation = trpc.viewer.me.updateProfile.useMutation({
+    onSuccess: async () => {
+      try {
+        // Create default event types if user has none
+        if (eventTypes?.length === 0) {
+          await Promise.all(
+            DEFAULT_EVENT_TYPES.map(async (event) => {
+              return createEventType.mutateAsync({
+                title: t(event.title),
+                slug: event.slug,
+                length: event.length,
+                hidden: event.hidden,
+              });
+            })
+          );
+        }
+      } catch (error) {
+        console.error(error);
+      }
+
+      await utils.viewer.me.get.refetch();
+      router.push("/event-types");
+    },
+    onError: (error) => {
+      showToast(t("something_went_wrong"), "error");
+      console.error(error);
+    },
+  });
+
+  const submitPersonalOnboarding = () => {
+    telemetry.event(telemetryEventTypes.onboardingFinished);
+    mutation.mutate({
+      completedOnboarding: true,
+    });
+  };
+
+  return {
+    submitPersonalOnboarding,
+    isSubmitting: mutation.isPending,
+  };
+};
